requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
sqlparse==0.4.3
whitenoise==6.4.0
